:recycle: Refactor translations and move packages
