+++ src/components/AutonomousAgent.ts
@@ -1,38 +1,29 @@
 import axios from "axios";
-import type { ModelSettings } from "../utils/types";
-import type { Analysis } from "../services/agent-service";
+import type {ModelSettings} from "../utils/types";
+import type {Analysis} from "../services/agent-service";
 import AgentService from "../services/agent-service";
+import {DEFAULT_MAX_LOOPS_CUSTOM_API_KEY, DEFAULT_MAX_LOOPS_FREE, DEFAULT_MAX_LOOPS_PAID,} from "../utils/constants";
+import type {Session} from "next-auth";
+import {env} from "../env/client.mjs";
+import {v1, v4} from "uuid";
+import type {RequestBody} from "../utils/interfaces";
+import type {AgentMode, AgentPlaybackControl, Message, Task,} from "../types/agentTypes";
 import {
-  DEFAULT_MAX_LOOPS_CUSTOM_API_KEY,
-  DEFAULT_MAX_LOOPS_FREE,
-  DEFAULT_MAX_LOOPS_PAID,
-} from "../utils/constants";
-import type { Session } from "next-auth";
-import { env } from "../env/client.mjs";
-import { v4, v1 } from "uuid";
-import type { RequestBody } from "../utils/interfaces";
-import {
-  AUTOMATIC_MODE,
-  PAUSE_MODE,
-  AGENT_PLAY,
   AGENT_PAUSE,
-  TASK_STATUS_STARTED,
-  TASK_STATUS_EXECUTING,
-  TASK_STATUS_COMPLETED,
-  TASK_STATUS_FINAL,
-  MESSAGE_TYPE_TASK,
+  AGENT_PLAY,
+  AUTOMATIC_MODE,
   MESSAGE_TYPE_GOAL,
-  MESSAGE_TYPE_THINKING,
   MESSAGE_TYPE_SYSTEM,
+  MESSAGE_TYPE_TASK,
+  MESSAGE_TYPE_THINKING,
+  PAUSE_MODE,
+  TASK_STATUS_COMPLETED,
+  TASK_STATUS_EXECUTING,
+  TASK_STATUS_FINAL,
+  TASK_STATUS_STARTED,
 } from "../types/agentTypes";
-import type {
-  AgentMode,
-  Message,
-  Task,
-  AgentPlaybackControl,
-} from "../types/agentTypes";
-import { useAgentStore, useMessageStore } from "./stores";
-import { i18n } from "next-i18next";
+import {useAgentStore, useMessageStore} from "../stores";
+import {translate} from "../utils/translations";
 
 const TIMEOUT_LONG = 1000;
 const TIMOUT_SHORT = 800;
@@ -191,11 +182,7 @@ class AutonomousAgent {
     } catch (e) {
       console.log(e);
       this.sendErrorMessage(
-        `${i18n?.t(
-          "ERROR_ADDING_ADDITIONAL_TASKS",
-          "ERROR_ADDING_ADDITIONAL_TASKS",
-          { ns: "errors" }
-        )}`
+        translate("ERROR_ADDING_ADDITIONAL_TASKS", "errors")
       );
 
       this.sendMessage({ ...currentTask, status: TASK_STATUS_FINAL });
@@ -203,9 +190,11 @@ class AutonomousAgent {
     await this.loop();
   }
 
-  getRemainingTasks() {
-    const tasks = useMessageStore.getState().tasks;
-    return tasks.filter((task: Task) => task.status === TASK_STATUS_STARTED);
+  getRemainingTasks(): Task[] {
+    return useMessageStore
+      .getState()
+      .tasks
+      .filter((t: Task) => t.status === TASK_STATUS_STARTED) as Task[];
   }
 
   private conditionalPause() {
@@ -338,10 +327,7 @@ class AutonomousAgent {
 
       if (axios.isAxiosError(e) && e.response?.status === 429) {
         this.sendErrorMessage(
-          `${i18n?.t("RATE_LIMIT_EXCEEDED", "RATE_LIMIT_EXCEEDED", {
-            ns: "errors",
-          })}`
-        );
+          translate("RATE_LIMIT_EXCEEDED", "errors"));
       }
 
       throw e;
@@ -380,33 +366,23 @@ class AutonomousAgent {
   sendLoopMessage() {
     this.sendMessage({
       type: MESSAGE_TYPE_SYSTEM,
-      value: !!this.modelSettings.customApiKey
-        ? `${i18n?.t("AGENT_MAXED_OUT_LOOPS", "AGENT_MAXED_OUT_LOOPS", {
-            ns: "errors",
-          })}`
-        : `${i18n?.t("DEMO_LOOPS_REACHED", "DEMO_LOOPS_REACHED", {
-            ns: "errors",
-          })}`,
+      value: translate(!!this.modelSettings.customApiKey
+        ? "AGENT_MAXED_OUT_LOOPS" : "DEMO_LOOPS_REACHED",
+        'errors')
     });
   }
 
   sendManualShutdownMessage() {
     this.sendMessage({
       type: MESSAGE_TYPE_SYSTEM,
-      value: `${i18n?.t(
-        "AGENT_MANUALLY_SHUT_DOWN",
-        "AGENT_MANUALLY_SHUT_DOWN",
-        { ns: "errors" }
-      )}`,
+      value: translate("AGENT_MANUALLY_SHUT_DOWN", "errors")
     });
   }
 
   sendCompletedMessage() {
     this.sendMessage({
       type: MESSAGE_TYPE_SYSTEM,
-      value: `${i18n?.t("ALL_TASKS_COMPLETETD", "ALL_TASKS_COMPLETETD", {
-        ns: "errors",
-      })}`,
+      value: translate("ALL_TASKS_COMPLETETD", "errors")
     });
   }
 
@@ -453,23 +429,15 @@ const testConnection = async (modelSettings: ModelSettings) => {
 };
 
 const getMessageFromError = (e: unknown) => {
-  let message = `${i18n?.t("ERROR_ACCESSING_OPENAI_API_KEY", {
-    ns: "errors",
-  })}`;
+  let message = "ERROR_RETRIEVE_INITIAL_TASKS";
+
   if (axios.isAxiosError(e)) {
-    const axiosError = e;
-    if (axiosError.response?.status === 429) {
-      message = `${i18n?.t("ERROR_API_KEY_QUOTA", {
-        ns: "errors",
-      })}`;
-    }
-    if (axiosError.response?.status === 404) {
-      message = `${i18n?.t("ERROR_OPENAI_API_KEY_NO_GPT4", { ns: "errors" })}`;
-    }
-  } else {
-    message = `${i18n?.t("ERROR_RETRIEVE_INITIAL_TASKS", { ns: "errors" })}`;
+    if (e.response?.status === 429) message = "ERROR_API_KEY_QUOTA"
+    if (e.response?.status === 404) message = "ERROR_OPENAI_API_KEY_NO_GPT4"
+    else message = "ERROR_ACCESSING_OPENAI_API_KEY"
   }
-  return message;
+
+  return translate(message, "errors");
 };
 
 export default AutonomousAgent;
