+++ src/stores/agentStore.ts
